backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
authlib==1.2.1
itsdangerous==2.1.2

